requirements.txt
# local package
-e .

# external requirements
click
python-dotenv>=0.5.1
gdown